Works I guess?

+++ mod/eldenring/allocator-hook/src/allocations.rs
@@ -1,9 +1,8 @@
 use std::ops;
 use std::sync;
 use std::collections;
-use std::collections::Bound::{Excluded, Included};
 
-static mut ALLOCATION_TABLE: Option<sync::RwLock<collections::BTreeMap<usize, usize>>> = None;
+static mut ALLOCATION_TABLE: Option<sync::RwLock<collections::BTreeMap<u64, u64>>> = None;
 
 pub fn init_allocation_table() {
     unsafe {
@@ -11,7 +10,7 @@ pub fn init_allocation_table() {
     }
 }
 
-pub fn register_allocation(ptr: usize, size: usize) {
+pub fn register_allocation(ptr: u64, size: u64) {
     let mut table = unsafe {
         ALLOCATION_TABLE.as_mut().unwrap().write().unwrap()
     };
@@ -19,7 +18,7 @@ pub fn register_allocation(ptr: usize, size: usize) {
     table.insert(ptr, size);
 }
 
-pub fn remove_allocation(ptr: usize) {
+pub fn remove_allocation(ptr: u64) {
     let mut table = unsafe {
         ALLOCATION_TABLE.as_mut().unwrap().write().unwrap()
     };
@@ -27,7 +26,7 @@ pub fn remove_allocation(ptr: usize) {
     table.remove(&ptr);
 }
 
-pub fn get_memory_page_range(ptr: usize) -> ops::Range<usize> {
+pub fn get_memory_page_range(ptr: u64) -> ops::Range<u64> {
     let nth_page = ptr / 4096;
     let lower = nth_page.clone() * 4096;
     let upper = (nth_page.clone() + 1) * 4096;
@@ -38,7 +37,7 @@ pub fn get_memory_page_range(ptr: usize) -> ops::Range<usize> {
     }
 }
 
-pub fn page_contains_allocation(ptr: usize) -> bool {
+pub fn page_contains_allocation(ptr: u64) -> bool {
     let page = get_memory_page_range(ptr);
     let table = unsafe {
         ALLOCATION_TABLE.as_ref().unwrap().read().unwrap()
